we now alway apply a median filter to segmented images before passing them down cgal mesh generator.
git-svn-id: https://nirfast-meshing.googlecode.com/svn/image2mesh_cgal/trunk@347 fb8e7f84-bfd6-7a22-e4a7-981ca257bf3e

+++ image2mesh_cgal.m
@@ -52,7 +52,7 @@ if isfield(param,'medfilter') && param.medfilter==1
     param.medfilter = 1;
 end
 [mask info] = GetImageStack(fn,param);
-mask = uint8(mask);
+mask = uint16(mask);
 
 if isfield(info,'Offset')
     param.Offset = info.Offset;
